updated: postgres

+++ cmd/web/components/toSlice.go
@@ -9,7 +9,7 @@ func ToSlice(s string) []string {
     return strings.Split(s, ",")
 }
 
-func NewURL(eventId int64) string {
+func NewURL(eventId int32) string {
    id := int(eventId) 
    return strings.Join([]string{"/event", strconv.Itoa(id)}, "-")
 }
